+++ src/http/_config.js
@@ -1,9 +1,8 @@
 let { join } = require('path')
 
-module.exports = function config (inv, cwd) {
+module.exports = function config ({ apigateway, cwd, inv }) {
   // Handle API type
-  let apiType = process.env.ARC_API_TYPE || inv.aws.apigateway || 'http'
-  delete process.env.ARC_API_TYPE // Destroy to purify testing
+  let apiType = process.env.ARC_API_TYPE || inv.aws.apigateway || apigateway || 'http'
 
   // Allow override of 'public' folder
   let staticPath = join(cwd, inv.static.folder)